Fix importer (#9)
The GTFS Schedule Importer job is failing due to changes and errors in the structure of hosted data. It should be able to process the data, or if there is a serious issue with the hosted data, that should be brought to the attention of the hosting transit agency.

Seeing duplicate stops in a single route. Notified the transit agency. Waiting for a response.
https://gist.github.com/s2t2/0d2929e0ecaba85823e1314935e7941e#gistcomment-2796526